Move conversion to nodes to a separate file
This commit creates a new file `nodes.jl` which contains the code that
converts input matrices into `SVMNode`s.

Moreover, the function that converts instances represented as a sparse
matrix is a bit changed to use only exported functions from
SparseArrays.

When the matrix containing kernel values is sparse, we convert it to a
full matrix and use the full matrix implementation. Hopefully, it should
not matter too much as there is a good chance that sparse matrices are
not that useful for kernel values.