Add support for precomputed kernel. (#73)
* Add support for precomputed kernel.

This commit is the necessary minimum to support precomputed kernel. I
tried to keep the number of changes low, so the code could probably use
some polishing. I also added a basic test (the expected values are based
on the output of `svm-train` from LIBSVM).

* Change `ArgumentError` to `DimensionMismatch`.

We now throw `DimensionMismatch` instead of `ArgumentError` in case of
invalid shape of the input matrix.

* Adjust `svmpredict` to support precomputed kernel.

The prediction with precomputed kernel now works. Although, the code
gets uglier and uglier.

* Add more difficult test case.

It also fixes getting the number of instances for prediction in the case
of precomputed kernel.

* Add a test case with Iris data.

* Add a test-case with a sparse input matrix

* Start indexing `SVMNode`s from 1

* Precomp. kernel: special method for creating nodes

In the case of precomputed kernel, the first column of nodes must
contain indices of the instances. This special method, `gram2nodes`,
adds the indices implicitly so we don't have to add them to the input
matrix explicitly.

* Unify conversion to nodes

A method is introduced to create `SVMNode`s from ordinary instances or
from a precomputed kernel matrix. So far, it only works for full
matrices, not sparse ones.

* Move conversion to nodes to a separate file

This commit creates a new file `nodes.jl` which contains the code that
converts input matrices into `SVMNode`s.

Moreover, the function that converts instances represented as a sparse
matrix is a bit changed to use only exported functions from
SparseArrays.

When the matrix containing kernel values is sparse, we convert it to a
full matrix and use the full matrix implementation. Hopefully, it should
not matter too much as there is a good chance that sparse matrices are
not that useful for kernel values.

* Add docs for precomputed kernel

* Use a single instance for all terminal `SVMNode`s

This commit introduces a constant instance of `SVMNode(-1, NaN)` that is
used as every terminal `SVMNode`.

* Remove unnecessary `using`

* Add input check for regression machines